Fixed issue with race condition on SQLAlchemy + PostgreSQL table create
