+++ packages/grid/apps/domain/src/main/core/database/__init__.py
@@ -10,6 +10,7 @@ db = SQLAlchemy()
 
 
 class BaseModel(db.Model, AllFeaturesMixin):
+    __bind_key__ = "bin_store"
     __abstract__ = True
     pass
 
@@ -19,7 +20,6 @@ from .bin_storage.bin_obj import BinObject
 from .bin_storage.bin_obj import ObjectMetadata
 from .bin_storage.json_obj import JsonObject
 from .bin_storage.metadata import StorageMetadata
-from .dataset.datasetgroup import DatasetGroup
 from .groups.groups import Group
 from .groups.usergroup import UserGroup
 from .requests.request import Request
