+++ packages/grid/apps/domain/src/main/core/database/bin_storage/bin_obj.py
@@ -40,6 +40,7 @@ class ObjectMetadata(BaseModel):
 
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     obj = db.Column(db.String(3072), db.ForeignKey("bin_object.id"))
+
     tags = db.Column(db.JSON())
     description = db.Column(db.String())
     read_permissions = db.Column(db.JSON())
