+++ packages/grid/apps/domain/src/main/core/datasets/dataset_ops.py
@@ -23,11 +23,11 @@ import torch as th
 from ..database import db
 from ..database.bin_storage.bin_obj import BinObject
 from ..database.bin_storage.bin_obj import ObjectMetadata
+from ..database.bin_storage.json_obj import BinObjDataset
+from ..database.bin_storage.json_obj import Dataset
+from ..database.bin_storage.json_obj import DatasetGroup
 from ..database.bin_storage.json_obj import JsonObject
 from ..database.bin_storage.metadata import get_metadata
-from ..database.dataset.datasetgroup import BinObjDataset
-from ..database.dataset.datasetgroup import Dataset
-from ..database.dataset.datasetgroup import DatasetGroup
 from ..database.store_disk import DiskObjectStore
 from ..database.utils import model_to_json
