+++ packages/grid/apps/network/src/main/core/database/bin_storage/json_obj.py
@@ -4,7 +4,6 @@ from .. import db
 
 
 class JsonObject(BaseModel):
-    __bind_key__ = "bin_store"
     __tablename__ = "json_object"
 
     id = db.Column(db.String(), primary_key=True)
