+++ packages/grid/apps/worker/src/main/core/database/bin_storage/bin_obj.py
@@ -4,7 +4,6 @@ from .. import db
 
 
 class BinaryObject(BaseModel):
-    __bind_key__ = "bin_store"
     __tablename__ = "binary_object"
 
     id = db.Column(db.String(), primary_key=True)
